Firebase Firestore SER-TR Dictionary App

+++ sozluk_app_ser_tr_fbfs/lib/screens/home_page.dart
@@ -1,4 +1,6 @@
 /// <----- home_page.dart ----->
+import 'dart:developer';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 
@@ -111,6 +113,7 @@ class _HomePageState extends State<HomePage> {
                 ),
         ],
       ),
+
       /// FAB Basılınca uygulanacak metot
       floatingActionButton: FloatingActionButton(
         onPressed: () => openWordBox(),
@@ -144,6 +147,7 @@ class _HomePageState extends State<HomePage> {
 
                 return Padding(
                   padding: const EdgeInsets.only(left: 10, right: 10),
+
                   /// kelime tıklanınca detay sayfası açılıyor
                   child: GestureDetector(
                     onTap: () {
@@ -208,7 +212,67 @@ class _HomePageState extends State<HomePage> {
                               ),
                               IconButton(
                                 onPressed: () {
-                                  firestoreService.deleteWord(word.wordId);
+                                  log("delete word ?");
+                                  showDialog(
+                                    context: context,
+                                    builder: (BuildContext context) {
+                                      return AlertDialog(
+                                        shape: RoundedRectangleBorder(
+                                          borderRadius:
+                                              BorderRadius.circular(16.0),
+                                          side: const BorderSide(
+                                            color: Colors.red,
+                                            width: 2.0,
+                                          ),
+                                        ),
+                                        title: const Text(
+                                          "Dikkat !!!",
+                                          style: TextStyle(
+                                            color: Colors.red,
+                                            fontWeight: FontWeight.bold,
+                                          ),
+                                        ),
+                                        content: Row(
+                                          children: [
+                                            const Text("Bu kelime "),
+                                            Text(
+                                              "(${word.sirpca})",
+                                              style: const TextStyle(
+                                                fontWeight: FontWeight.bold,
+                                                color: Colors.blueAccent,
+                                                fontSize: 16,
+                                              ),
+                                            ),
+                                            const Text(" silinsin mi ?"),
+                                          ],
+                                        ),
+                                        actions: [
+                                          TextButton(
+                                            child: const Text("İptal"),
+                                            onPressed: () {
+                                              log("İptal seçildi");
+                                              Navigator.pop(context);
+                                            },
+                                          ),
+                                          TextButton(
+                                            child: const Text(
+                                              "Tamam",
+                                              style: TextStyle(
+                                                fontWeight: FontWeight.bold,
+                                              ),
+                                            ),
+                                            onPressed: () {
+                                              log("Tamam seçildi");
+                                              firestoreService.deleteWord(word.wordId);
+                                              Navigator.pop(context);
+                                            },
+                                          ),
+                                        ],
+                                      );
+                                    },
+                                  );
+
+                                  /// firestoreService.deleteWord(word.wordId);
                                 },
                                 icon: const Icon(Icons.delete),
                                 tooltip: "kelime sil",
